chore: refactor VTP controller.
