+++ src/logion/controllers/locrequest.controller.ts
@@ -71,7 +71,6 @@ export function fillInSpec(spec: OpenAPIV3.Document): void {
     LocRequestController.submitLocRequest(spec);
     LocRequestController.cancelLocRequest(spec);
     LocRequestController.reworkLocRequest(spec);
-    LocRequestController.setVerifiedThirdParty(spec);
 }
 
 type CreateLocRequestView = components["schemas"]["CreateLocRequestView"];
@@ -87,7 +86,6 @@ type AddFileView = components["schemas"]["AddFileView"];
 type AddLinkView = components["schemas"]["AddLinkView"];
 type AddMetadataView = components["schemas"]["AddMetadataView"];
 type CreateSofRequestView = components["schemas"]["CreateSofRequestView"];
-type SetVerifiedThirdPartyRequest = components["schemas"]["SetVerifiedThirdPartyRequest"];
 
 export type UserPrivateData = {
     identityLocId: string | undefined,
@@ -929,31 +927,5 @@ export class LocRequestController extends ApiController {
                 walletUser: userIdentity,
             }
         }
-    }
-
-    static setVerifiedThirdParty(spec: OpenAPIV3.Document) {
-        const operationObject = spec.paths["/api/loc-request/{requestId}/verified-third-party"].put!;
-        operationObject.summary = "Sets the VTP flag of the closed Identity LOC";
-        operationObject.description = "The authenticated user must be the owner of the LOC.";
-        operationObject.responses = getDefaultResponsesNoContent();
-        operationObject.requestBody = getRequestBody({
-            description: "VTP flag",
-            view: "SetVerifiedThirdPartyRequest",
-        });
-        setPathParameters(operationObject, {
-            'requestId': "The ID of the LOC",
-        });
-    }
-
-    @HttpPut('/:requestId/verified-third-party')
-    @Async()
-    @SendsResponse()
-    async setVerifiedThirdParty(body: SetVerifiedThirdPartyRequest, requestId: string) {
-        const request = requireDefined(await this.locRequestRepository.findById(requestId));
-        const userCheck = await this.authenticationService.authenticatedUser(this.request);
-        userCheck.require(user => user.is(request.ownerAddress));
-        request.setVerifiedThirdParty(body.isVerifiedThirdParty || false);
-        await this.locRequestRepository.save(request);
-        this.response.sendStatus(204);
     }
 }
